refactor(racingcar): Move race logic to new executor class
Seperate car racing logic from read input
